✨ feat: user input handling with single function
